+++ src/main/java/chess/model/piece/Bishop.java
@@ -1,5 +1,7 @@
 package chess.model.piece;
 
+import chess.model.position.Movement;
+
 public class Bishop extends Piece {
     private static final Piece BLACK_BISHOP = new Bishop(Color.BLACK);
     private static final Piece WHITE_BISHOP = new Bishop(Color.WHITE);
