+++ src/main/java/chess/model/piece/Pawn.java
@@ -1,5 +1,7 @@
 package chess.model.piece;
 
+import chess.model.position.Movement;
+
 public class Pawn extends Piece {
     private static final Piece WHITE_PAWN = new Pawn(Color.WHITE, 1, 2);
     private static final Piece BLACK_PAWN = new Pawn(Color.BLACK, -1, 7);
